Apply `PageHeader` and `Breadcrumbs` component from `@cartridge/ui` (#126)
