--- src/modules/Home/components/Stats.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Stats() {
-  return <div>Stats</div>;
-}
